+++ packages/react/src/components/AnimatedHeader/components/Tiles/BaseTile/BaseTile.tsx
@@ -6,7 +6,7 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import React, { ReactNode } from 'react';
+import React, { ElementType, ReactNode } from 'react';
 import { AIPromptTile } from '../AIPromptTile/AIPromptTile';
 import { GlassTile } from '../GlassTile/GlassTile';
 
@@ -15,46 +15,23 @@ import { GlassTile } from '../GlassTile/GlassTile';
 interface BaseTileProps {
   id?: string;
   open?: boolean;
-  href?: string;
-  mainIcon?: string;
-  secondaryIcon?: string;
-  title?: string;
-  subtitle?: string;
+  href?: string | null;
+  mainIcon?: ElementType | null;
+  secondaryIcon?: ElementType | null;
+  title?: string | null;
+  subtitle?: string | null;
   productName?: string;
   customContent?: ReactNode;
-  onClick?: () => void;
+  isLoading?: boolean;
+  isDisabled?: boolean;
+  disabledTaskLabel?: string;
+  onClick?: (() => void) | null;
 }
 
-export const BaseTile: React.FC<BaseTileProps> = ({
-  id,
-  open,
-  href,
-  mainIcon,
-  secondaryIcon,
-  title,
-  subtitle,
-  productName,
-  customContent,
-  onClick,
-}: BaseTileProps) => {
-  const props = {
-    id,
-    open,
-    href,
-    mainIcon,
-    secondaryIcon,
-    title,
-    subtitle,
-    productName,
-    customContent,
-    onClick,
-  };
-  const tile =
-    id === 'ai-tile' ? (
-      <AIPromptTile {...props}></AIPromptTile>
-    ) : (
-      <GlassTile {...props}></GlassTile>
-    );
+export const BaseTile: React.FC<BaseTileProps> = (props: BaseTileProps) => {
+  if (props.id === 'ai-tile') {
+    return <AIPromptTile {...props}></AIPromptTile>;
+  }
 
-  return tile;
+  return <GlassTile {...props}></GlassTile>;
 };